+++ src/functions/build.ts
@@ -1,7 +1,7 @@
 import { normalize } from "@zakahacecosas/string-utils";
 import { Commander } from "./cli.ts";
 import { LogStuff } from "./io.ts";
-import { FknError } from "./error.ts";
+import { DebugFknErr, FknError } from "./error.ts";
 
 export function RunBuildCmds(commands: string[]) {
     for (const command of commands) {
@@ -20,17 +20,19 @@ export function RunBuildCmds(commands: string[]) {
                 ],
             );
             if (!out.success) {
-                throw new FknError(
+                LogStuff(out.stdout ?? "(No stdout/stderr was written by the command)");
+                DebugFknErr(
                     "Task__Build",
-                    `Command "${command}" has failed (command #${cmdIndex} in your 'buildCmd' sequence). We've halted execution. Scroll up, as output of this command (if any) should appear in this terminal session.`,
+                    `Command "${command}" has failed (command #${cmdIndex} in your 'buildCmd' sequence). We've halted execution.`,
+                    out.stdout,
+                    false,
                 );
             }
             if (normalize(out.stdout).length === 0) LogStuff("No output received.", undefined, ["half-opaque", "italic"]);
             else LogStuff(out.stdout);
             LogStuff("Done!", undefined, "bold");
         } catch (error) {
-            LogStuff(String(error));
-            LogStuff("Something went wrong...", "danger", "red");
+            LogStuff((error as FknError).message, "danger", "red");
             // halt execution, especially to avoid releases
             Deno.exit(1);
         }